prevent replacing *-orig files that cause recursive calls when install script is run twice

+++ install.sh
@@ -45,8 +45,16 @@ chmod +x /opt/xensource/sm/RBDSR.py
 chmod +x /opt/xensource/sm/cephutils.py
 ln -s /opt/xensource/sm/RBDSR.py /opt/xensource/sm/RBDSR
 
-mv /sbin/tap-ctl /sbin/tap-ctl-orig
-mv /bin/vhd-tool /bin/vhd-tool-orig
+if [ -e /sbin/tap-ctl-orig ] then
+  echo "tap-ctl-orig already in place, not backing up!"
+else
+  mv /sbin/tap-ctl /sbin/tap-ctl-orig
+fi
+if [ -e /bin/vhd-tool-orig ] then
+  echo "/bin/vhd-tool-orig already in place, not backing up!"
+else
+  mv /bin/vhd-tool /bin/vhd-tool-orig
+fi
 cp bins/tap-ctl /sbin/tap-ctl
 cp bins/vhd-tool /bin/vhd-tool
 chmod +x /sbin/tap-ctl
